avoid compiler warning (cast from int to float)

+++ contrib/brl/bseg/sdet/sdet_vehicle_finder.cxx
@@ -75,7 +75,7 @@ n_regions_closest_to_pick(vcl_vector<sdet_region_sptr> const& regions,
         continue;
       float x0 = (*rit)->Xo(), y0 = (*rit)->Yo();
       int xp = pick_.x(), yp = pick_.y();
-      float d = vcl_sqrt((x0-xp)*(x0-xp) + (y0-yp)*(y0-yp));
+      float d = (float)vcl_sqrt((x0-xp)*(x0-xp) + (y0-yp)*(y0-yp));
       double area = (*rit)->area();
       float max_d = distance_scale_*vcl_sqrt(area);
 #ifdef DEBUG
